requirements.txt
django
Pillow
whitenoise
#mysqlclient
django-ckeditor
python-decouple
django-email-verification